MP3: Invent, use "audiosample_tock"
.. as in "tick, tock".  A cause of skipping in mp3 playback is that
to fill an output buffer, sometimes SD needs to be read.  This takes
at least 7ms, which can use up most of the time budget available to
do the decoding and whatever other steps prepare the audio buffer for
output.

Instead, think of the time to decode the mp3 as the "tick", after which
it is made available to the output device; the new "tock" refills the
input buffer just when we know there's the greatest amount of buffered
decoded audio available.