MP3Decoder: rework "single channel" handling
What I didn't understand was that initially, the sequence of channel
numbers is "0011", while in normal operation it should consistently
be "0101".  Just keeping track of the "other channel" works during the
steady state, but at startup it causes the right channel to be shifted
by 1.

This cures the small glitches that could be seen clearly when playing
stereo sine wave mp3s.