audiosample: Remove single_channel and channel parameters
These are no longer needed, as the implementation that ever passed
'true' to single_channel is gone.

This also allows removal of some state from audio sample sources,
such as the "channel read count" of WaveFile.

This change reindents code and as such may be better to view with
a whitespace-ignoring diff such as "git show -w"