framebufferdisplay: introduce put_pixel_span to framebuffer protocol
It would be nice to apply framebufferdisplay to smart LED strips arranged
in grids.

Based on conversations with @krayola, these are almost always not laid
out linearly in memory.  Instead, one or both of these techniques are
used:
 - zig zag, so that e.g., row 0 goes from left to right, and then row 1
   goes from right to left, returning to where it started
 - multi-string setups like with neopxl8

Additionally, to work with the _neopixel library, the pre- and post-fix
data need to be avoided.

All these things mean that the underlying implementer of the framebuffer
protocol needs to be able to customize how the pixel values get stored in
memory.

So, this patch introduces a new "put_pixel_span" method in the protocol,
and provides an implementation that does exactly what the original inlined
implementation did.  Implementations for zigzag pixel strings will follow.