framebufferdisplay: introduce can_survive_reset to framebuffer protocol
The requirement that all FramebufferDisplays be able to survive reset is
too strict.  At a minimum, it means that Python code can never be used
to implement one.  Additionally, due to small size, some displays are not
useful for the console, and the only way to use a display after reset that
is NOT hard-coded into the CircuitPython build is as a console.
(i.e., board.DISPLAY exists)

As written, this is a per-type flag, rather than a getter function. For
now that suffices, but it can be upgraded to a function if warranted.