tools: Allow specific boards to be built "monolingual"
We have several translations which only fit on resource-constrained
boards due to being incomplete.  With contributions on weblate,
some builds no longer fit.

Rather than rendering ourselves unable to accept updated weblate
translations, this change allows specific boards to be switched to
monolingual.  It also adds the boards which are failing in the latest
weblate PR due to space to the list. (#2942)