Further tweak makeqstrdata
This unfortunately doesn't help with the verbosest language of the moment,
German!  It shrinks a bit off zh_Latn_pinyin but that's not the problem.
It actually makes German worse despite the attempts to carefully find
an "optimum" dictionary size.

The C decoder hasn't properly been taught about the 9-bit mode, so this
won't run anyway.