MP3Decoder: When assigning the "file", close the old one
JP ran into problems with playing multiple MP3 files within the same
program.  One stumbling point was around the MP3Decoder.file property.

Before this, it was the responsibility of the user to close the old
file, which led to inconvenient and non-obvious code such as
    old_file = decoder.file
    decoder.file = open("example.mp3", "w")
    old_file.close()

Compatibility with this idiom is preserved, as it's permitted to close()
a file as many times as desired; the second and subsequent close()s do
nothing.