canio: RemoteTransmissionRequest: make it its own C type
This puts tendrils into send and receive that needlessly obscure and bloat
the code.  There are now two copies of the send code (one for each
C structure type, each in their own function) and two copies of the
receive code (one for each C structure type, but both in the single
'read' function).