Proof of concept: Let files be built as C++
.. this requires changing all the port Makefiles, so it's only been
done for samd so far.

Trickery is needed to permit the CMSIS headers be included.

All headers need the MP_BEGIN_DECLS / MP_END_DECLS macro added.

The <stdatomic.h> header is not compatible with the C++ compiler, so
it is hacked out.

The file built as C++ is shared-modules/util.cc and it doesn't actually
use any real C++ features such as classes or templates.

runtime type inference and exceptions are disabled.