translations: Allow a slightly larger word list
This saves 76 bytes overall in metro_m0_express de_DE, at the cost of
increasing the number of CPU cycles to decompress a word.

Before this, an array called "wends" tracked the Word ENDS of each
word.  This limited the total size of the word array in bytes to a bit
more than 256 (because extra knowledge that each word was at least 2
bytes long was used to stretch the effective range of the 8 bit values).

Now, the length of each word is stored.  Since these lengths are small,
they fit within bytes.  The cost is, you have to count through the list
of wlens (Word LENgthS).

In fact, further savings appears possible: sort words by length; store
only an 8 element array giving the number of words of each possible length
from 2 to 9 inclusive.  This could save over 200 additional bytes!  It's
similar to how the huffman code lengths are stored.

Testing performed: that en_US and de_DE language messages were correctly
printed in the unix port.  The `help()` message was printed incorrectly
when I had a bug.