UNTESTED audiocore: Simplify code when only a single DMA channel used
For some DAC configurations on SAM D5x/E5x, two DMA channels are required.
This adds extra bookkeeping that is not needed on SAM D21 (or for any other
current microcontroller families).  This saves
about 100 bytes of flash storage.  It also saves a modest amount
of RAM (12 bytes per WaveFile object)

The same optimization could be applied to MP3, but hasn't been yet since
it's not on any devices that are so resource-constrained that 100 bytes
helps.