PulseIn: add frequency selection
.. implemented only on raspberrypi, where it's easiest.  If this is
considered useful enough, we can potentially implement it on at least SOME
other ports.  However, each port is a bit different so for now I just made them
ignore the frequency= constructor parameter and return 1_000_000 for the
frequency property.

This is useful to me because using a frequency of 1000 lets me track the length
of WWWB symbols, which range in length from 200ms to 800ms long.