shared runtime: Improve buffer handling
This simplifies the pair of calls mp_get_buffer_raise + normalize_buffer_bounds
into a single call. It also lays the groundwork for the element
size to be used in normalization, so that e.g., if you
work with an array of 16-bit values, `write(buf, start=3, end=9)`
will work in 2 byte chunks, just like `write(buf[3:9])`.

However, the latter part of the change is not enabled yet,
it's just that the place it would need to be done is centralized
by this change.

This change also saves 104 bytes on trinket m0, which is nice.