Don't send escape codes during raw REPL
In the raw REPL, all usual output (such as echoing back characters)
is stopped. This should probably include the status bar, since otherwise
there's no way for a GUI like Thonny to differentiate between output
from the code it executes remotely, vs the status bar changing.

(This is separate from whether the regular friendly repl's status bar
updates will confuse Thonny; in that case, it'll be up to Thonny to show
or suppress the status bar updates as appropriate)