Reduce the size of the unicode case conversion table
This saves hundreds of bytes of flash, at the expense of no longer
performing correct case-insensitive comparisons of filenames that
contain obscure characters, on non-FULL_BUILDs.

The table still covers all characters in Microsoft Glyph List 4,
(Latin, Central European, Cyrillic, Greek, Turkish, Baltic, and
CP437). However, it does represent a regression for filenames in
"obscure" languages.

An example of a character pair that is no longer compared correctly
is from the IPA (international phonetic alphabet) section:
```
Ɓ \N{LATIN CAPITAL LETTER B WITH HOOK} U+0181
ɓ \N{LATIN SMALL LETTER B WITH HOOK} U+0253
```

A workaround for any problem that arises from this would simply to code
the filename with the same capitalization in the code as on the filesystem
(i.e., the files are still usable, but they just are functionally
case sensitive, unlike 'CODE.PY' vs 'code.py')

Note that these letters (as well as many of the letters still covered by
the abbreviated table) are already not correctly case-converted
when using the Python upper() and lower() methods in CircuitPython:
```
Adafruit CircuitPython 8.0.0-beta.1-41-g55519670a8-dirty on 2022-10-13; Raspberry Pi Pico W with rp2040
>>> "Ɓ".lower()
'Ɓ'
>>> "ɓ".upper()
'ɓ'
```