Extend error checking of str, bytes encoding and errors arguments
I'm not sure we _want_ to do this, it trades memory for no new
functionality except more exceptions.

Tests cases are changed because "ignore" and "replace" were never
supported and because the utf-8 encoding should always be specified by
name, rather than depending on the empty string.

(However, the empty string is accepted as a valid encoding name for
compatibility with previous versions of CircuitPython)