DOESN'T WORK and would break OnDiskBitmap if it did...
Expand displayio transactions

In principle, we would like to use background SPI DMA so that pixels
can be sent while new pixels are being calculated. This could potentially
reduce frame time from ~252ms to ~185ms in my test (3.9fps to 5.4fps).

However, a number of things stand in the way of this. In order for a
background write to be useful, the bus has to remain owned by the displayio
core. To end a transaction, the last background write has to finish.
Thus, initial efforts didn't bear any fruit since immediately after
every data chunk was sent, the transaction was ended.

On the other hand, there's at least one difficult-to-change reason that
it is this way: OnDiskBitmap. OnDiskBitmap is already a problem for
bus locking and performance, but also the OnDiskBitmap _could_ be on
the same SPI bus as the display, so you can't enter displayio_display_core_fill_area
with the bus lock held.

So .. our hands are tied by the system of displayio as it currently
exists. There appears to be no way to incrementally evolve it to overlap
pixel calculation & display bus writing.

And anyway this patch as it is does not work, it never even initializes
and clears the LCD.