KeyMatrix: support treating row pins as address pins.
Two keyboards I own have a key matrix setup, except that the outputs are
a set of address-select pins.

While various sources use different conventions, CircuitPython has made
the choice that the *row* pins are selecting pins and the *column* pins
are reading pins, so this PR follows that scheme by converting *row*
pins to *address* pins.

This can differ from the nomenclature used in original documentation or
reverse engineered documentation, which may use either "row" or "column"
nomenclature for the addressed pins.

It's possible that fewer than all address values are valid.

Example 1. Atari XEGS keyboard with 6 address pins + 2 data pins
    - All addresses are valid
    - (though 2nd data pin value is used only to scan the control &
      break keys so only 2 of its 64 positions are useful!)
    - This example would pass "True" to the new keyword argument.

Example 2. Unicomp Mini M keyboard with 5 address pins + 12 data pins (16 addresses valid)
    - Two of the pins are chip select lines of 8-to-1 multiplexers, so
      they always have opposite values (one chip selected at a time)
    - This example would pass a list of 16 ints to the new keyword argument.

Both keyboards happen to not have per-key diodes, but there does not
seem to be any reason why this HAS to be the case.

This adds support for an "address-scanning" mode of KeyMatrix. I haven't
tested it, but this feature should support both of those keyboard
styles as well as others.