Implement the official E9 workaround
& enable it on all RP2350 builds.

This is sort of OK for slow polling from Python code, as the input buffer
is almost always disabled even if the `.value` property of a DigitalInOut
is retrieved in a tight Python loop.

But is not useful for touchio: touchio must continuously sample the pin,
which causes its input buffer to be on almost all the time, negating the
workaround.

in my test setup trying to make touchio work, the ~18ns the input
buffer is enabled is almost enough to replace the charge lost in ~20us
of discharge via a 3.3Mohm through hole resistor to GND,
with the only extra capacitance being my scope probe.

And it is not a workaround for PIO, as PIO cannot control the input
buffer enable.