audiocore: remove support for "single buffer"
This is only needed on samd51, and most of the complexity can be
implemented just once instead of in each audio sample type.

This saves 968 bytes flash on feather rp2350. More importantly,
it'll save some fraction as much on samd21 where stereo audio output
is not supported at all (but savings will be less overall because
fewer audiosample types are available)

Other builds besides raspberrypi are broken & raspberrypi is not even
tested.