py: Add, use MP_RETURNS_TWICE modifier.
Like setjmp, nlr_push returns twice. The compiler needs to know
this.

This annotation makes gcc do "setjmp-like" things in order to
make that work.

It might also show new diagnostics about clobbered
variables. For instance, there was such a warning in nativeglue
because the `send_value` automatic variable was not qualified
`volatile` and was changed after the nlr_push invocation.
(C99 standard 7.13.2.1.3)

In that case, rewriting the code so that a new `to_send` variable
with a smaller scope is the one that is modified, prevented the
clobber diagnostic.

Signed-off-by: Jeff Epler <jepler@gmail.com>