nativeglue: Avoid clobber diagnostic.
Because `send_value` has automatic storage, when it is
mutated, its value can not be assumed after the call to `nlr_push`.
This analysis is spurious (since the second return of nlr_push is
to the else branch, and send_value is not used in that arm) but
it occurs when nlr_push is marked as returns_twice (including when
setjmp is used to implement nlr).

Signed-off-by: Jeff Epler <jepler@gmail.com>