core: Fix mp_printf integer size mismatches.
The type of the argument must match the format string. Add
casts and/or modify format strings in order to ensure that they do.

These locations were found using an experimental gcc plugin
for mp_printf error checking, cross-building for x64 windows
on Linux.

In one case there was already a cast, but it was written
incorrectly and did not have the intended effect.

Signed-off-by: Jeff Epler <jepler@gmail.com>