ports: Add "m68int16"
This builds a Linux binary that works under qemu-m68k. It operates
in "-mshort" mode, in which ints are 16 bit but pointers & longs
are 32 bit.

It uses a toolchain built from https://github.com/kentosama/m68k-elf-gcc
which is pretty old. It's not packaged up in any way that's easy
to use from CI.

It doesn't do file I/O yet, and so I haven't run the testsuite
against it.c