mpprint: Correctly format leading zeros with separators.
In this alternate version, the spaces/zeroes arrays are removed
to save code size. Instead, a 4-byte buffer is prepared on
stack for any non-grouping fill byte.

This likely improves non-zero, non-space padding perf a little,
reduces zero and space padding perf a little, and saves some bytes
of flash.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>